requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
aiosmtplib==1.1.6
alembic==1.11.1
amqp==5.3.1
anyio==4.9.0
asyncpg==0.30.0
attrs==25.3.0
billiard==4.2.1
boto3==1.39.4
botocore==1.39.4
celery==5.3.1
certifi==2025.7.9
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cryptography==45.0.5
databases==0.9.0
dataclasses-json==0.5.14
dnspython==2.7.0
email-validator==1.3.1
fastapi==0.95.2
filelock==3.18.0
frozenlist==1.7.0
fsspec==2025.5.1
greenlet==3.2.3
grpcio==1.74.0
h11==0.14.0
h2==4.2.0
hf-xet==1.1.5
hpack==4.1.0
httpcore==0.17.3
httptools==0.6.4
httpx==0.24.1
huggingface-hub==0.33.2
hyperframe==6.1.0
idna==3.10
IMAPClient==2.3.0
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.1
kombu==5.5.4
langchain==0.0.201
langchainplus-sdk==0.0.20
lxml==6.0.0
Mako==1.3.10
MarkupSafe==3.0.2
marshmallow==3.26.1
mpmath==1.3.0
multidict==6.6.3
mypy_extensions==1.1.0
networkx==3.5
nltk==3.9.1
numexpr==2.11.0
numpy==1.26.4
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
openapi-schema-pydantic==1.2.4
packaging==25.0
pdfminer.six==20250506
pillow==11.3.0
portalocker==3.2.0
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==6.31.1
psycopg2-binary==2.9.6
pycparser==2.22
pydantic==1.10.11
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.0.0
python-multipart==0.0.20
PyYAML==6.0.2
qdrant-client==1.15.0
redis==4.6.0
regex==2024.11.6
requests==2.32.4
s3transfer==0.13.0
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.16.0
sentence-transformers==2.2.2
sentencepiece==0.2.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.19
starlette==0.27.0
sympy==1.14.0
tenacity==8.5.0
threadpoolctl==3.6.0
tokenizers==0.13.3
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
transformers==4.33.2
triton==3.3.1
typing-inspect==0.9.0
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.22.0
uvloop==0.21.0
vine==5.1.0
watchfiles==1.1.0
wcwidth==0.2.13
websockets==15.0.1
yarl==1.20.1
